moves dinolistitem component into own folder
